techpack: audio: Configure for awinic and foursemi
awinic has two different configurations depending if we are building for
rhode or not, so let's just set it up to use our cflag for it and properly
setup the makefiles too.

Since we're cleaning up awinic? We might as well do the same for foursemi
as well and also setup the Kconfig where it will get read.

Change-Id: If49d560a82a95e7417ae8668663736fa3fb69458